third_party/WebKit/LayoutTests/storage/indexeddb/deletedatabase-delayed-by-open-and-versionchange-expected.txt

Test that deleteDatabase is delayed if a VERSION_CHANGE transaction is running

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


dbname = "deletedatabase-delayed-by-open-and-versionchange.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)

prepareDatabase():
sawVersionChange = false
upgradeTransactionComplete = false
h = event.target.result
indexedDB.deleteDatabase(dbname)

onVersionChange():
PASS event.target.version is 1
PASS event.oldVersion is 1
PASS event.newVersion is null
sawVersionChange = true
Connection is not closed, so 'blocked' should fire

transactionOnComplete():
upgradeTransactionComplete = true

onOpenSuccess():
h = event.target.result
h.close()
Closing too late to prevent the in-flight 'blocked' event

deleteDatabaseOnBlocked():
PASS sawVersionChange is true

deleteDatabaseOnSuccess():
PASS upgradeTransactionComplete is true
PASS successfullyParsed is true

TEST COMPLETE
